Add CSS File to Bower `main`
Automated build tools like [wiredep](https://www.npmjs.com/package/wiredep) work better when CSS files are included as part of the `main` configuration in the bower.json file. The bower.json spec also indicates that styles should be included if they should be used in conjunction with the file.