Comments Added
In this PR, the dragging container css is changed such that whenever we drag any image to the container it will show the `drag and drop` popup and when we drag out the image from that container this popup will disappear.
The thing which will look better after this:- Presently, when we drag the image anywhere in the editor it will display a popup of `drag and drop` which is not a good UI or experience to have because we have another drag and drop too which is MainImage Module. After this PR the popup will only appear we dargin the image over that particular container.